First iteration of removing legacy image library

+++ app/src/main/java/org/telegram/android/util/IOUtils.java
@@ -14,8 +14,10 @@ public class IOUtils {
     };
 
     public static void delete(File src) {
-        if (!src.delete()) {
-            src.deleteOnExit();
+        if (src.exists()) {
+            if (!src.delete()) {
+                src.deleteOnExit();
+            }
         }
     }
